[edit]
1. fragment로 작업한 걸 activity로 변경중
