+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".CreateRoomActivity"
+            android:name=".CreateRoomLocalActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
