+++ app/src/main/java/com/example/testapplication/MainActivity.kt
@@ -1,9 +1,11 @@
 package com.example.testapplication
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentContainerView
+import com.example.testapplication.databinding.ActivityMainBinding
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
@@ -14,16 +16,25 @@ class MainActivity : AppCompatActivity() {
     private val bottomNagivationView: BottomNavigationView by lazy { // 하단 네비게이션 바
         findViewById(R.id.bnv_main)
     }
-
+    private lateinit var binding1: ActivityMainBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+        binding1 = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding1.root)
+
+        binding1.btnTestCreate.setOnClickListener{
+            val intent= Intent(this, CreateRoomLocalActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
 
         if (frame == null) {
             supportFragmentManager.beginTransaction()
                 .add(frame.id, RoomaddFragment())
                 .commit()
         }
+
         bottomNagivationView.setOnNavigationItemSelectedListener {item ->
             when(item.itemId) {
                 R.id.nav_category -> {
