+++ app/src/main/res/layout/fragment_selectoption.xml
@@ -16,7 +16,7 @@
         android:fontFamily="@font/pretendard_bold"
         android:text="방탈출 옵션 선택"
         android:textColor="@color/text_main"
-        android:textSize="22sp"
+        android:textSize="24sp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -39,7 +39,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="100dp"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.553"
+        app:layout_constraintHorizontal_bias="0.861"
         app:layout_constraintStart_toEndOf="@+id/thema"
         app:layout_constraintTop_toBottomOf="@+id/textView7" />
 
@@ -63,6 +63,7 @@
         android:layout_marginTop="20dp"
 
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.85"
         app:layout_constraintStart_toEndOf="@+id/difficulty"
         app:layout_constraintTop_toBottomOf="@+id/thema" />
 
@@ -86,6 +87,7 @@
         android:layout_marginTop="20dp"
 
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.85"
         app:layout_constraintStart_toEndOf="@+id/scary"
         app:layout_constraintTop_toBottomOf="@+id/difficulty" />
